refactor test infrastructure, unify pytest workflow, and fix DDS test process cleanup
