+++ .dockerignore
@@ -1,27 +1,57 @@
+# ---------------------------------------------------------------------------
 # Build artifacts and caches
+# ---------------------------------------------------------------------------
 build/
-.venv*/
+**/build/
 __pycache__/
+.pytest_cache/
+.ruff_cache/
+.rumdl_cache/
 *.pyc
 
-# IDE
+# ---------------------------------------------------------------------------
+# Virtual environments
+# ---------------------------------------------------------------------------
+.venv*/
+
+# ---------------------------------------------------------------------------
+# IDE and OS metadata
+# ---------------------------------------------------------------------------
 .vscode/
 .idea/
+.DS_Store
+Thumbs.db
 
-# Git
+# ---------------------------------------------------------------------------
+# Docker context exclusions (VCS and CI metadata)
+# ---------------------------------------------------------------------------
 .git/
 .github/
 
-# Security artifacts (generated at build time)
+# ---------------------------------------------------------------------------
+# Generated source artifacts
+# ---------------------------------------------------------------------------
+Types.py
+builtin_logging_type.py
+ThreatTypes.py
+*_timestamp.cmake
+
+# ---------------------------------------------------------------------------
+# Security and runtime artifacts
+# ---------------------------------------------------------------------------
+rti_license.dat
 *.pem
 *.p7s
 *.csr
 *.srl
 *.psk
 
-# Recordings
+# ---------------------------------------------------------------------------
+# Application runtime output
+# ---------------------------------------------------------------------------
 or_recording/
 
-# OS files
-.DS_Store
-Thumbs.db
+# ---------------------------------------------------------------------------
+# Misc temporary files
+# ---------------------------------------------------------------------------
+*.glade~
