+++ .gitignore
@@ -1,17 +1,51 @@
-.vscode/
-.venv*/
+# ---------------------------------------------------------------------------
+# Build artifacts and caches
+# ---------------------------------------------------------------------------
 build/
+**/build/
 __pycache__/
+.pytest_cache/
+.ruff_cache/
+.rumdl_cache/
+*.pyc
+
+# ---------------------------------------------------------------------------
+# Virtual environments
+# ---------------------------------------------------------------------------
+.venv*/
+
+# ---------------------------------------------------------------------------
+# IDE and OS metadata
+# ---------------------------------------------------------------------------
+.vscode/
+.idea/
+.DS_Store
+Thumbs.db
+
+# ---------------------------------------------------------------------------
+# Generated source artifacts
+# ---------------------------------------------------------------------------
 Types.py
 builtin_logging_type.py
 ThreatTypes.py
 *_timestamp.cmake
-*.pyc
+
+# ---------------------------------------------------------------------------
+# Security and runtime artifacts
+# ---------------------------------------------------------------------------
 rti_license.dat
 *.pem
 *.p7s
 *.csr
-*.glade~
 *.srl
 *.psk
+
+# ---------------------------------------------------------------------------
+# Application runtime output
+# ---------------------------------------------------------------------------
 or_recording/
+
+# ---------------------------------------------------------------------------
+# Misc temporary files
+# ---------------------------------------------------------------------------
+*.glade~
