--- .markdownlint.json
@@ -1,12 +1,0 @@
-{
-  "MD013": false,
-  "MD033": {
-    "allowed_elements": ["details", "summary", "strong"]
-  },
-  "MD024": {
-    "siblings_only": true
-  },
-  "MD060": {
-    "style": "compact"
-  }
-}
