+++ modules/04-security-threat/security/.gitignore
@@ -1,7 +1,12 @@
-# Generated security artifacts — re-run setup_threat_security.py to regenerate.
+# ---------------------------------------------------------------------------
+# Generated security artifacts
+# ---------------------------------------------------------------------------
+# Re-run setup_threat_security.py to regenerate.
 # Committed files: *.cnf, *.xml (unsigned), setup_threat_security.py.
 
+# ---------------------------------------------------------------------------
 # CA state and generated files (keys, certs, CRLs)
+# ---------------------------------------------------------------------------
 ca/*/index.txt
 ca/*/index.txt.*
 ca/*/serial
@@ -11,12 +16,18 @@ ca/*/certs/
 ca/*/newcerts/
 ca/*/requests/
 
-# Signed governance and permissions (.p7s)
+# ---------------------------------------------------------------------------
+# Signed governance and permissions
+# ---------------------------------------------------------------------------
 domain_scope/**/signed/
 
+# ---------------------------------------------------------------------------
 # Identity keys and certificates
+# ---------------------------------------------------------------------------
 identity/**/private/
 identity/**/certs/
 
+# ---------------------------------------------------------------------------
 # Scaffold hash sidecar
+# ---------------------------------------------------------------------------
 .scaffold-hash
