+++ tests/README.md
@@ -1,82 +1,138 @@
 # Project-level Tests
 
-Tests that validate the project-wide build pipeline and cross-module
-concerns. Module-specific tests live under each module's own `tests/`
-directory.
+Tests that validate the project-wide build pipeline and cross-module concerns.
+Module-specific tests live under each module's own `tests/` directory.
 
 ## Prerequisites
 
 | Requirement | Notes |
 | --- | --- |
-| RTI Connext DDS 7.3.0+ | `NDDSHOME` and `CONNEXTDDS_ARCH` must be set |
-| Python 3.9+ | With `rti.connextdds` and `pytest` installed |
+| RTI Connext DDS 7.3.1+ | `NDDSHOME` and `CONNEXTDDS_ARCH` must be set (via rtisetenv script) |
+| Python 3.9+ | With `rti.connextdds` and `pytest` installed (in `.venv39` or equivalent) |
 | C++ build complete | Run `python build.py` first |
 
-## Running Tests
+## Quick Start
 
-From the repository root:
+From the repository root, after sourcing the RTI environment and activating the Python venv:
 
 ```bash
-# All project-level tests (includes cmake re-configure)
-python -m pytest tests/ -v
+# Run all 130 tests (repo-level + all modules)
+python -m pytest
 
-# Skip cmake configure/build (just verify binaries exist)
-python -m pytest tests/ -v -k "not cmake"
+# Run with verbose output
+python -m pytest -v
 
-# Markdown lint (rumdl)
-rumdl check .
+# Run specific module or test
+python -m pytest modules/01-operating-room/tests/test_types.py -v
 
-# Full project test sweep (repo-level + module-level tests)
-tests/run_tests.sh -v
+# Run tests matching a pattern
+python -m pytest -k "test_types" -v
 
-# Full project test sweep without rebuilding
-tests/run_tests.sh -v --skip-build
-
-# Full project test sweep in containerized environment
-docker run --rm \
-  -v "$RTI_LICENSE_FILE:/opt/rti.com/rti_connext_dds-7.3.1/rti_license.dat:ro" \
-  medtech-test
+# Run a specific test class or function
+python -m pytest tests/test_config_parsing.py::TestModuleJsonContract -v
 ```
 
-## Choosing a Test Path
+## Environment Setup
 
-Use this quick guide to pick the right command:
+The root `pyproject.toml` configures pytest to automatically discover tests in:
 
-| Goal | Preferred approach | Why |
-| --- | --- | --- |
-| Fast local iteration on one test/file | `pytest` directly | Fastest feedback loop and easiest targeted debugging |
-| Validate only project-level tests under `tests/` | `pytest tests/ -v` | Runs just the root test suite defined in this folder |
-| Validate end-to-end test coverage before merge (repo + modules) | `tests/run_tests.sh` | Runs root tests and each module's `tests/` suite in one command |
-| Reproduce CI/container-specific behavior or dependency issues | Docker test image | Ensures clean, repeatable environment close to pipeline execution |
+- `tests/` — Project-level tests
+- `modules/01-operating-room/tests/` — Module 01 tests
+- `modules/02-record-playback/tests/` — Module 02 tests
+- `modules/04-security-threat/tests/` — Module 04 tests
 
-### When to Prefer `tests/run_tests.sh`
+Before running pytest, ensure the RTI environment is sourced:
 
-Prefer `tests/run_tests.sh` when you need confidence across the whole repository, not just root tests. In particular:
+```bash
+source /opt/rti.com/rti_connext_dds-7.3.1/resource/scripts/rtisetenv_x64Linux4gcc7.3.0.bash
+source .venv39/bin/activate
+```
 
-1. Before pushing or opening a PR, to catch module-level failures that `pytest tests/` does not execute.
-2. After changing shared code used by multiple modules.
-3. After build or environment changes where a full sweep is safer than targeted checks.
+## Test Running Approaches
 
-Use `--skip-build` if binaries are already up to date and you only want to rerun tests quickly.
+### Direct pytest (Recommended for Day-to-Day)
 
-### When to Prefer Explicit `pytest`
+Use root-level `pytest` directly for fast iteration and clear output:
 
-Prefer direct `pytest` commands during day-to-day development when you are focused on one area and want fast feedback:
+```bash
+# All tests
+python -m pytest
 
-1. Running a single test file/class/function while debugging.
-2. Working only on repo-level tests under `tests/`.
-3. Tight edit-test cycles where full project sweeps would be too slow.
+# With markers to skip slow/gui/secure tests
+python -m pytest -m "not slow and not gui and not secure"
 
-### When to Prefer Docker Testing
+# Fast feedback on a single file
+python -m pytest tests/test_config_parsing.py -v
+```
 
-Prefer the Docker path when environment parity matters more than speed:
+**When to use:**
 
-1. Reproducing failures that do not occur on your host machine.
-2. Verifying behavior with containerized dependencies/tooling.
-3. Final validation for environments where local host differences may hide issues.
+- Daily development on single tests or modules
+- Tight edit-test cycles where speed matters
+- Focused debugging of one area
+
+### Docker Container Testing
+
+Use Docker for reproducible, isolated test environments:
+
+```bash
+# Set up license file
+export RTI_LICENSE_FILE=/path/to/rti_license.dat
+
+# Run all tests in container
+docker compose -f tests/docker/docker-compose.yml run --rm --build test
+
+# Run specific tests
+docker compose -f tests/docker/docker-compose.yml run --rm --build test \
+    -v modules/01-operating-room/tests/test_types.py
+
+# Run tests matching a pattern
+docker compose -f tests/docker/docker-compose.yml run --rm --build test \
+    -k "test_types"
+```
+
+**When to use:**
+
+- Reproducing CI failures in a clean environment
+- Verifying behavior when host dependencies vary
+- Final pre-merge validation
+- Testing GUI components on headless systems (Xvfb included)
+
+### Markdown and Other Checks
+
+```bash
+# Markdown linting (rumdl)
+rumdl check .
+
+# Code formatting and linting (pre-commit)
+pre-commit run --all-files
+```
+
+## Test Markers
+
+Tests can be filtered using pytest markers. Common markers:
+
+- `@pytest.mark.slow` — Long-running tests (e.g., full application demos)
+- `@pytest.mark.gui` — GUI tests requiring a display
+- `@pytest.mark.secure` — Security tests requiring artifacts and plugins
+
+Run fast tests only (skip slow/GUI/secure):
+
+```bash
+python -m pytest -m "not slow and not gui and not secure"
+```
+
+Run only GUI tests:
+
+```bash
+python -m pytest -m "gui"
+```
 
 ## Test Structure
 
 | File | What it tests |
 | --- | --- |
-| `test_build.py` | CMake configure & build succeed; Module 01 C++ binaries exist |
+| `test_project_build_pipeline.py` | CMake configure & build; Module 01 C++ binaries & shared libraries exist |
+| `test_config_parsing.py` | Module JSON schemas and configuration parsing for all modules |
+| `test_security_status.py` | Security artifacts generation and availability for secure tests |
+| `test_markdown_lint.py` | Project documentation (README, Scenario, etc.) passes rumdl checks |
