+++ tests/docker/README.md
@@ -30,12 +30,20 @@ export RTI_LICENSE_FILE=/path/to/rti_license.dat
 ### Using Docker Compose (recommended)
 
 ```bash
-# One command: build image (if needed), run tests, and remove container
+# Run all tests (default: verbose output)
 docker compose -f tests/docker/docker-compose.yml run --rm --build test
 
-# Run specific tests
+# Run specific test file with custom pytest args
 docker compose -f tests/docker/docker-compose.yml run --rm --build test \
-    python -m pytest modules/01-operating-room/tests/test_types.py -v
+    modules/01-operating-room/tests/test_types.py -v
+
+# Run tests matching a pattern
+docker compose -f tests/docker/docker-compose.yml run --rm --build test \
+    -k "test_types"
+
+# Run fast tests only (skip slow/GUI/secure)
+docker compose -f tests/docker/docker-compose.yml run --rm --build test \
+    -m "not slow and not gui and not secure"
 
 # Optional: compose up style (then clean up service container)
 docker compose -f tests/docker/docker-compose.yml up --build --abort-on-container-exit --exit-code-from test test
@@ -47,11 +55,11 @@ docker compose -f tests/docker/docker-compose.yml down --remove-orphans
 ```bash
 # From the repo root:
 
-# 1. Build
+# 1. Build both images
 docker build -f tests/docker/Dockerfile --target build -t medtech-build .
 docker build -f tests/docker/Dockerfile --target test  -t medtech-test .
 
-# 2. Run all tests (mount license at runtime)
+# 2. Run all tests (mounts license, default pytest with -v)
 docker run --rm \
     -v $RTI_LICENSE_FILE:/opt/rti.com/rti_connext_dds-7.3.1/rti_license.dat:ro \
     medtech-test
@@ -60,13 +68,32 @@ docker run --rm \
 docker run --rm \
     -v $RTI_LICENSE_FILE:/opt/rti.com/rti_connext_dds-7.3.1/rti_license.dat:ro \
     medtech-test \
-    python -m pytest modules/01-operating-room/tests/test_types.py -v
+    modules/01-operating-room/tests/test_types.py -v
 
-# 4. Interactive shell for debugging
+# 4. Run tests matching a pattern
+docker run --rm \
+    -v $RTI_LICENSE_FILE:/opt/rti.com/rti_connext_dds-7.3.1/rti_license.dat:ro \
+    medtech-test \
+    -k "test_types"
+
+# 5. Interactive shell for debugging
 docker run --rm -it \
     -v $RTI_LICENSE_FILE:/opt/rti.com/rti_connext_dds-7.3.1/rti_license.dat:ro \
     medtech-test bash
 ```
+
+## Understanding the Test Flow
+
+The Docker entrypoint (`tests/docker/entrypoint.sh`):
+
+1. Starts Xvfb virtual display (`:99`) for headless GUI tests
+2. Sources the RTI Connext environment
+3. Runs `python -m pytest` with any provided arguments
+4. Cleans up Xvfb on exit
+
+Example: `docker run medtech-test` → entrypoint calls `python -m pytest -v`
+
+Example: `docker run medtech-test -k "test_types"` → entrypoint calls `python -m pytest -k "test_types"`
 
 ## Why two images?
 
@@ -84,3 +111,5 @@ docker run --rm -it \
   build stage so secure tests can run.
 - `docker compose run --rm ...` avoids leftover one-off containers after test
   completion.
+- All 130 tests (repo-level + modules) are collected and run via the unified
+  `python -m pytest` approach configured in the root `pyproject.toml`.
