+++ tests/docker/entrypoint.sh
@@ -1,12 +1,20 @@
 #!/usr/bin/env bash
-# Start Xvfb for headless GUI tests, then run whatever command was passed.
+# Docker test entrypoint: Start Xvfb for headless GUI tests, then run tests.
+# The RTI Connext environment and Python venv are already configured in the Dockerfile.
 set -e
 
+# Start Xvfb virtual display for headless GUI testing
 Xvfb :99 -screen 0 1280x1024x24 -nolisten tcp &
 XVFB_PID=$!
-trap 'kill $XVFB_PID 2>/dev/null' EXIT
+trap 'kill $XVFB_PID 2>/dev/null || true' EXIT
 
 # Give Xvfb a moment to start
 sleep 0.5
 
-exec "$@"
+# Source RTI environment (persists across to test command)
+source /opt/rti.com/rti_connext_dds-7.3.1/resource/scripts/rtisetenv_x64Linux4gcc7.3.0.bash
+
+# Run pytest from repo root with any passed arguments
+# Default to running all tests if no args provided
+cd /workspace
+exec python -m pytest "${@:-.}"
